更正分布式系统中的文件传输方式
支持分布式大文件数据以的流的形式式传输

+++ member-dao/src/test/java/dao/MemberDaoTest.java
@@ -20,7 +20,7 @@ public class MemberDaoTest {
     @Test
     public void test(){
         MemberExample e = new MemberExample();
-        int i  = memberMapper.countByExample(e);
+        int i  = (int) memberMapper.countByExample(e);
         Member member = new Member();
         System.out.println(i);
     }
